Day 13 - Part 1 (use int64)

+++ 13/main.go
@@ -8,12 +8,12 @@ import (
 )
 
 type machineInfo struct {
-	buttonAX int
-	buttonAY int
-	buttonBX int
-	buttonBY int
-	prizeX   int
-	prizeY   int
+	buttonAX int64
+	buttonAY int64
+	buttonBX int64
+	buttonBY int64
+	prizeX   int64
+	prizeY   int64
 }
 
 func panicErr(e error) {
@@ -37,21 +37,21 @@ func main() {
 			coordY, e2 := strconv.Atoi(coords[1][2:])
 			panicErr(e2)
 			if lineParts[0] == "Button A" {
-				machine.buttonAX = coordX
-				machine.buttonAY = coordY
+				machine.buttonAX = int64(coordX)
+				machine.buttonAY = int64(coordY)
 			} else if lineParts[0] == "Button B" {
-				machine.buttonBX = coordX
-				machine.buttonBY = coordY
+				machine.buttonBX = int64(coordX)
+				machine.buttonBY = int64(coordY)
 			} else if lineParts[0] == "Prize" {
-				machine.prizeX = coordX
-				machine.prizeY = coordY
+				machine.prizeX = int64(coordX)
+				machine.prizeY = int64(coordY)
 			}
 		}
 
 		machines = append(machines, machine)
 	}
 
-	totalCost := 0
+	totalCost := int64(0)
 	for i, machine := range machines {
 		a, b := solveEquation(machine.buttonAX, machine.buttonAY, machine.buttonBX, machine.buttonBY, machine.prizeX, machine.prizeY)
 		fmt.Printf("Machine %d - (%d, %d) + (%d, %d) = (%d, %d) has solution (%d, %d)\n", i, machine.buttonAX, machine.buttonAY, machine.buttonBX, machine.buttonBY, machine.prizeX, machine.prizeY, a, b)
@@ -66,7 +66,7 @@ func main() {
 	fmt.Println(totalCost)
 }
 
-func solveEquation(ax, ay, bx, by, px, py int) (int, int) {
+func solveEquation(ax, ay, bx, by, px, py int64) (int64, int64) {
 	// Solve system of equations:
 	// ax * t + bx * s = px
 	// ay * t + by * s = py
